Create PC11. Ch13 Payroll.cpp
Design a payroll class that has data members for an employees hourly pay rate, number of hours worked, and total pay for the week. Write a program with an array of seven payroll objects. The program should ask the user for the number of hours each employee has worked and will then display the amount of gross pay each has earned.
input validation on hours worked to 60.